+++ src/RepetierServerSharpApi/RepetierClient.cs
@@ -678,41 +678,6 @@ namespace AndreasReitberger.API.Repetier
         }
         #endregion
 
-        #region Timers
-        void StopPingTimer()
-        {
-            if (PingTimer != null)
-            {
-                try
-                {
-                    PingTimer?.Change(Timeout.Infinite, Timeout.Infinite);
-                    PingTimer = null;
-                    IsListeningToWebsocket = false;
-                }
-                catch (ObjectDisposedException)
-                {
-                    //PingTimer = null;
-                }
-            }
-        }
-        void StopTimer()
-        {
-            if (Timer != null)
-            {
-                try
-                {
-                    Timer?.Change(Timeout.Infinite, Timeout.Infinite);
-                    Timer = null;
-                    IsListening = false;
-                }
-                catch (ObjectDisposedException)
-                {
-                    //PingTimer = null;
-                }
-            }
-        }
-        #endregion
-
         #endregion
 
         #region Public
@@ -925,7 +890,7 @@ namespace AndreasReitberger.API.Repetier
         #endregion
 
         #region ActivePrinter
-        public new async Task SetPrinterActiveAsync(int index = -1, bool refreshPrinterList = true)
+        public override async Task SetPrinterActiveAsync(int index = -1, bool refreshPrinterList = true)
         {
             try
             {
@@ -955,7 +920,7 @@ namespace AndreasReitberger.API.Repetier
             }
         }
         
-        public new async Task SetPrinterActiveAsync(string slug, bool refreshPrinterList = true)     
+        public override async Task SetPrinterActiveAsync(string slug, bool refreshPrinterList = true)     
         {
             try
             {
@@ -970,90 +935,7 @@ namespace AndreasReitberger.API.Repetier
                 OnError(new UnhandledExceptionEventArgs(exc, false));
             }
         }
-        
-        #endregion
 
-        #region WebCam
-        public string GetWebCamUri(int camIndex = 0, RepetierWebcamType type = RepetierWebcamType.Dynamic)
-        {
-            try
-            {
-                string currentPrinter = GetActivePrinterSlug();
-                if (string.IsNullOrEmpty(currentPrinter)) return string.Empty;
-
-                return $"{FullWebAddress}/printer/{(type == RepetierWebcamType.Dynamic ? "cammjpg" : "camjpg")}/{currentPrinter}?cam={camIndex}&apikey={ApiKey}";
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return "";
-            }
-        }
-        public async Task<string> GetWebCamUriAsync(int camIndex = 0, RepetierWebcamType type = RepetierWebcamType.Dynamic)
-        {
-            try
-            {
-                string currentPrinter = GetActivePrinterSlug();
-                if (string.IsNullOrEmpty(currentPrinter)) return string.Empty;
-
-                await RefreshPrinterConfigAsync();
-                return $"{FullWebAddress}/printer/{(type == RepetierWebcamType.Dynamic ? "cammjpg" : "camjpg")}/{currentPrinter}?cam={camIndex}&apikey={ApiKey}";
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return "";
-            }
-        }
-        public async Task<IWebCamConfig> GetWebCamConfigAsync(int camIndex = 0, bool refreshConfigs = true)
-        {
-            try
-            {
-                string currentPrinter = GetActivePrinterSlug();
-                if (string.IsNullOrEmpty(currentPrinter)) return null;
-
-                if (refreshConfigs)
-                {
-                    await RefreshPrinterConfigAsync();
-                }
-                
-                return WebCams?.FirstOrDefault(webCam => webCam.Position == camIndex);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return null;
-            }
-        }
-
-        public new Task<ObservableCollection<IWebCamConfig>> GetWebCamConfigsAsync() => GetWebCamConfigsAsync(GetActivePrinterSlug());
-
-        public async Task<ObservableCollection<IWebCamConfig>> GetWebCamConfigsAsync(string slug)
-        {
-            try
-            {
-                RepetierPrinterConfig config = await GetPrinterConfigAsync(slug);             
-                return new(config?.Webcams ?? new());
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return null;
-            }
-        }
-        public async Task<IWebCamConfig> GetWebCamConfigAsync(string slug, int camIndex = 0)
-        {
-            try
-            {
-                RepetierPrinterConfig config = await GetPrinterConfigAsync(slug);             
-                return config?.Webcams?.FirstOrDefault(webCam => webCam.Position == camIndex);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return null;
-            }
-        }
         #endregion
 
         #region CheckOnline
@@ -1382,540 +1264,6 @@ namespace AndreasReitberger.API.Repetier
             }
         }
 
-        #endregion
-
-        #region Models
-
-        public new Task GetFilesAsync() => GetModelsAsync(GetActivePrinterSlug(), GcodeImageType.Thumbnail, null);
-
-        public async Task<ObservableCollection<IGcode>> GetModelsAsync(
-            string PrinterName = "",
-            GcodeImageType ImageType = GcodeImageType.Thumbnail,
-            IProgress<int> Prog = null)
-        {
-            try
-            {
-                ObservableCollection<IGcode> modelDatas = new();
-                if (!IsReady)
-                    return modelDatas;
-
-                string currentPrinter = string.IsNullOrEmpty(PrinterName) ? GetActivePrinterSlug() : PrinterName;
-                if (string.IsNullOrEmpty(currentPrinter)) return modelDatas;
-
-                // Reporting
-                Prog?.Report(0);
-                RepetierModelList models = await GetModelListInfoResponeAsync(currentPrinter).ConfigureAwait(false);
-                if (models != null)
-                {
-                    List<RepetierModel> modelList = models.Data;
-                    if (modelList != null)
-                    {
-                        ObservableCollection<IGcode> Models = new(modelList);
-                        if (ImageType != GcodeImageType.None)
-                        {
-                            int total = Models.Count;
-                            for (int i = 0; i < total; i++)
-                            {
-                                IGcode model = Models[i];
-                                model.PrinterName = currentPrinter;
-                                model.ImageType = ImageType;
-                                // Load image depending on settings
-                                switch (ImageType)
-                                {
-                                    // Blocks thread, however async download leads to bad requestes
-                                    case GcodeImageType.Thumbnail:
-                                        model.Thumbnail = await GetDynamicRenderImageAsync(model.Identifier, true).ConfigureAwait(false);
-                                        break;
-                                    case GcodeImageType.Image:
-                                        model.Image = await GetDynamicRenderImageAsync(model.Identifier, false).ConfigureAwait(false);
-                                        break;
-                                    default:
-                                        model.Thumbnail = await GetDynamicRenderImageAsync(model.Identifier, true).ConfigureAwait(false);
-                                        model.Image = await GetDynamicRenderImageAsync(model.Identifier, false).ConfigureAwait(false);
-                                        break;
-                                }
-
-                                if (Prog != null)
-                                {
-                                    float progress = ((float)i / total) * 100f;
-                                    if (i < total - 1)
-                                    {
-                                        Prog.Report(Convert.ToInt32(progress));
-                                    }
-                                    else
-                                    {
-                                        Prog.Report(100);
-                                    }
-                                }
-                            }
-                        }
-                        else
-                        {
-                            Prog?.Report(100);
-                        }
-                        return Models;
-                    }
-                }
-
-                Prog?.Report(100);
-
-                return modelDatas;
-            }
-            catch (Exception exc)
-            {
-                if (Prog != null)
-                    Prog.Report(100);
-
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return new ObservableCollection<IGcode>();
-            }
-        }
-        
-        public async Task<Dictionary<long, byte[]>> GetModelImagesAsync(ObservableCollection<IGcode> models, GcodeImageType imageType = GcodeImageType.Thumbnail)
-        {
-            string currentPrinter = GetActivePrinterSlug();
-            if (string.IsNullOrEmpty(currentPrinter)) return null;
-
-            Dictionary<long, byte[]> result = new();
-            try
-            {
-                for (int i = 0; i < models.Count; i++)
-                {
-                    IGcode model = models[i];
-                    byte[] image = new byte[0];
-                    image = imageType switch
-                    {
-                        GcodeImageType.Thumbnail => await GetDynamicRenderImageAsync(model.Identifier, true).ConfigureAwait(false),
-                        GcodeImageType.Image => await GetDynamicRenderImageAsync(model.Identifier, false).ConfigureAwait(false),
-                        _ => throw new NotSupportedException($"The image type '{imageType}' is not supported here."),
-                    };
-                    result.Add(model.Identifier, image);
-                }
-                return result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return null;
-            }
-        }
-        public async Task<bool> DeleteModelFromServerAsync(IGcode model)
-        {
-            string currentPrinter = GetActivePrinterSlug();
-            if (string.IsNullOrEmpty(currentPrinter)) return false;
-
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{currentPrinter}";
-                IRestApiRequestRespone result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "removeModel",
-                       jsonObject: new { id = model.Identifier },
-                       authHeaders: AuthHeaders
-                       )
-                    .ConfigureAwait(false);
-                return GetQueryResult(result.Result);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-
-        public async Task RefreshModelsAsync(GcodeImageType imageType = GcodeImageType.Thumbnail, IProgress<int> prog = null)
-        {
-            try
-            {
-                ObservableCollection<IGcode> modelDatas = new();
-                if (!IsReady || ActivePrinter == null)
-                {
-                    Files = modelDatas;
-                    return;
-                }
-                Files = await GetModelsAsync(GetActivePrinterSlug(), imageType, prog).ConfigureAwait(false);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                Files = new ObservableCollection<IGcode>();
-            }
-        }
-        public async Task<bool> CopyModelToPrintQueueAsync(IGcode model, bool startPrintIfPossible = true)
-        {
-            string currentPrinter = GetActivePrinterSlug();
-            if (string.IsNullOrEmpty(currentPrinter))
-            {
-                return false;
-            }
-
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{currentPrinter}";
-                IRestApiRequestRespone result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "removeModel",
-                       jsonObject: new { id = model.Identifier, autostart = (startPrintIfPossible ? "true" : "false")},
-                       authHeaders: AuthHeaders
-                       )
-                    .ConfigureAwait(false);
-                /*
-                RepetierApiRequestRespone result =
-                    await SendRestApiRequestAsync(
-                        RepetierCommandBase.printer, RepetierCommandFeature.api,
-                        command: "copyModel", jsonData: $"{{\"id\":{model.Identifier}, \"autostart\":{(startPrintIfPossible ? "true" : "false")}}}",
-                        printerName: currentPrinter)
-                    .ConfigureAwait(false);
-                */
-                await RefreshJobListAsync().ConfigureAwait(false);
-                return GetQueryResult(result.Result, true);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-
-        public async Task UpdateFreeSpaceAsync()
-        {
-            IRestApiRequestRespone result = null;
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{GetActivePrinterSlug()}";
-                result = await SendRestApiRequestAsync(
-                   requestTargetUri: targetUri,
-                   method: Method.Post,
-                   command: "freeSpace",
-                   authHeaders: AuthHeaders
-                   )
-                .ConfigureAwait(false);
-
-                RepetierFreeSpaceRespone space = GetObjectFromJson<RepetierFreeSpaceRespone>(result.Result);
-                if (space != null)
-                {
-                    FreeDiskSpace = space.Free;
-                    TotalDiskSpace = space.Capacity;
-                    UsedDiskSpace = TotalDiskSpace - space.Available;
-                }
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(RepetierFreeSpaceRespone),
-                    Message = jecx.Message,
-                });
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-            }
-        }
-        public async Task RefreshDiskSpaceAsync()
-        {
-            await UpdateFreeSpaceAsync().ConfigureAwait(false);
-        }
-
-        #endregion
-
-        #region ModelGroups
-        public async Task<bool> AddModelGroupAsync(IGcodeGroup group) => await AddModelGroupAsync(group?.Name);
-
-        public async Task<bool> AddModelGroupAsync(string groupName)
-        {
-            string currentPrinter = GetActivePrinterSlug();
-            if (string.IsNullOrEmpty(currentPrinter)) return false;
-
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{currentPrinter}";
-                object data = new
-                {
-                    groupName = groupName,
-                };
-                IRestApiRequestRespone result = await SendRestApiRequestAsync(
-                   requestTargetUri: targetUri,
-                   method: Method.Post,
-                   command: "addModelGroup",
-                   jsonObject: data,
-                   authHeaders: AuthHeaders
-                   )
-                .ConfigureAwait(false);
-                /*
-                RepetierApiRequestRespone result =
-                    await SendRestApiRequestAsync(
-                        RepetierCommandBase.printer, RepetierCommandFeature.api,
-                        command: "addModelGroup", jsonData: string.Format("{{\"groupName\":\"{0}\"}}", groupName),
-                        printerName: currentPrinter)
-                    .ConfigureAwait(false);
-                */
-                return GetQueryResult(result.Result);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-        public async Task<bool> AddModelGroupAsync(string printerName, IGcodeGroup group) => await AddModelGroupAsync(printerName, group.Name);
-        public async Task<bool> AddModelGroupAsync(string printerName, string groupName)
-        {
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{printerName}";
-                object data = new
-                {
-                    groupName = groupName,
-                };
-                IRestApiRequestRespone result = await SendRestApiRequestAsync(
-                   requestTargetUri: targetUri,
-                   method: Method.Post,
-                   command: "addModelGroup",
-                   jsonObject: data,
-                   authHeaders: AuthHeaders
-                   )
-                .ConfigureAwait(false);
-                /*
-                RepetierApiRequestRespone result =
-                    await SendRestApiRequestAsync(
-                        RepetierCommandBase.printer, RepetierCommandFeature.api,
-                        command: "addModelGroup", jsonData: string.Format("{{\"groupName\":\"{0}\"}}", groupName),
-                        printerName: printerName)
-                    .ConfigureAwait(false);
-                */
-                return GetQueryResult(result.Result);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-
-        public async Task<bool> RemoveModelGroupAsync(IGcodeGroup group) => await RemoveModelGroupAsync(group.Name);
-        public async Task<bool> RemoveModelGroupAsync(string groupName)
-        {
-            string currentPrinter = GetActivePrinterSlug();
-            if (string.IsNullOrEmpty(currentPrinter)) return false;
-
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{currentPrinter}";
-                object data = new
-                {
-                    groupName = groupName,
-                };
-                IRestApiRequestRespone result = await SendRestApiRequestAsync(
-                   requestTargetUri: targetUri,
-                   method: Method.Post,
-                   command: "delModelGroup",
-                   jsonObject: data,
-                   authHeaders: AuthHeaders
-                   )
-                .ConfigureAwait(false);
-                /*
-                RepetierApiRequestRespone result =
-                    await SendRestApiRequestAsync(
-                        RepetierCommandBase.printer, RepetierCommandFeature.api,
-                        command: "delModelGroup", jsonData: string.Format("{{\"groupName\":\"{0}\"}}", groupName),
-                        printerName: currentPrinter)
-                    .ConfigureAwait(false);
-                */
-                return GetQueryResult(result.Result);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-        
-        public async Task<bool> RemoveModelGroupAsync(string printerName, IGcodeGroup group) => await RemoveModelGroupAsync(printerName, group.Name);
-        public async Task<bool> RemoveModelGroupAsync(string printerName, string groupName)
-        {
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{printerName}";
-                object data = new
-                {
-                    groupName = groupName,
-                };
-                IRestApiRequestRespone result = await SendRestApiRequestAsync(
-                   requestTargetUri: targetUri,
-                   method: Method.Post,
-                   command: "delModelGroup",
-                   jsonObject: data,
-                   authHeaders: AuthHeaders
-                   )
-                .ConfigureAwait(false);
-                /*
-                RepetierApiRequestRespone result =
-                    await SendRestApiRequestAsync(
-                        RepetierCommandBase.printer, RepetierCommandFeature.api,
-                        command: "delModelGroup", jsonData: string.Format("{{\"groupName\":\"{0}\"}}", groupName),
-                        printerName: printerName)
-                    .ConfigureAwait(false);
-                */
-                return GetQueryResult(result.Result);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-
-        public async Task<bool> MoveModelToGroupAsync(string groupName, IGcode file) => await MoveModelToGroupAsync(groupName, file.Identifier);
-        public async Task<bool> MoveModelToGroupAsync(string groupName, long id)
-        {
-            string currentPrinter = GetActivePrinterSlug();
-            if (string.IsNullOrEmpty(currentPrinter)) return false;
-
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{currentPrinter}";
-                object data = new
-                {
-                    groupName = groupName,
-                    id = id,
-                };
-                IRestApiRequestRespone result = await SendRestApiRequestAsync(
-                   requestTargetUri: targetUri,
-                   method: Method.Post,
-                   command: "moveModelFileToGroup",
-                   jsonObject: data,
-                   authHeaders: AuthHeaders
-                   )
-                .ConfigureAwait(false);
-                /*
-                RepetierApiRequestRespone result =
-                    await SendRestApiRequestAsync(
-                        RepetierCommandBase.printer, RepetierCommandFeature.api,
-                        command: "moveModelFileToGroup", jsonData: string.Format("{{\"groupName\":\"{0}\", \"id\":{1}}}", groupName, id),
-                        printerName: currentPrinter
-                        )
-                    .ConfigureAwait(false);
-                */
-                return GetQueryResult(result.Result);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-        public async Task<bool> MoveModelToGroupAsync(string printerName, IGcodeGroup group, IGcode file) => await MoveModelToGroupAsync(printerName, group.Name, file.Identifier);  
-        public async Task<bool> MoveModelToGroupAsync(string printerName, string groupName, long id)
-        {
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{GetActivePrinterSlug()}";
-                IRestApiRequestRespone result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                        command: "removeModel",
-                       jsonObject: new { groupName = groupName, id = id },
-                       authHeaders: AuthHeaders
-                       )
-                    .ConfigureAwait(false);
-                /*
-                RepetierApiRequestRespone result =
-                    await SendRestApiRequestAsync(
-                        RepetierCommandBase.printer, RepetierCommandFeature.api,
-                        command: "moveModelFileToGroup", jsonData: string.Format("{{\"groupName\":\"{0}\", \"id\":{1}}}", groupName, id),
-                        printerName: printerName
-                        )
-                    .ConfigureAwait(false);
-                */
-                return GetQueryResult(result.Result);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-
-        public async Task<ObservableCollection<IGcodeGroup>> GetModelGroupsAsync()
-        {
-            IRestApiRequestRespone result = null;
-            ObservableCollection<IGcodeGroup> resultObject = new();
-
-            string currentPrinter = GetActivePrinterSlug();
-            if (string.IsNullOrEmpty(currentPrinter)) return resultObject;
-
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{currentPrinter}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "listModelGroups",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result = await SendRestApiRequestAsync(
-                   commandBase: RepetierCommandBase.printer,
-                   commandFeature: RepetierCommandFeature.api, 
-                   command: "listModelGroups", 
-                   printerName: currentPrinter)
-                    .ConfigureAwait(false);
-                */
-                RepetierModelGroups info = GetObjectFromJson<RepetierModelGroups>(result.Result);
-                return info != null && info.GroupNames != null ? new ObservableCollection<IGcodeGroup>(info.GroupNames.Select(g => new RepetierModelGroup() { Name = g})) : resultObject;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(String),
-                    Message = jecx.Message,
-                });
-                return new ObservableCollection<IGcodeGroup>();
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshModelGroupsAsync()
-        {
-            try
-            {
-                ObservableCollection<IGcodeGroup> groups = new();
-                if (!IsReady || ActivePrinter == null)
-                {
-                    Groups = groups;
-                    return;
-                }
-
-                string currentPrinter = ActivePrinter.Slug;
-                if (string.IsNullOrEmpty(currentPrinter)) return;
-
-                RepetierModelGroups result = await GetModelGroupsAsync(currentPrinter).ConfigureAwait(false);
-                if (result != null)
-                {
-                    Groups = new ObservableCollection<IGcodeGroup>(result.GroupNames?.Select(g => new RepetierModelGroup() { Name = g }));
-                }
-                else Groups = groups;
-
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                Groups = new ObservableCollection<IGcodeGroup>();
-            }
-        }
         #endregion
 
         #region Jobs
@@ -2325,104 +1673,6 @@ namespace AndreasReitberger.API.Repetier
 
         #endregion
 
-        #region Printers
-        public new async Task<ObservableCollection<IPrinter3d>> GetPrintersAsync()
-        {
-            IRestApiRequestRespone result = null;
-            try
-            {
-                ObservableCollection<IPrinter3d> repetierPrinterList = new();
-                if (!IsReady)
-                    return repetierPrinterList;
-
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.List}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result = await SendRestApiRequestAsync(
-                    RepetierCommandBase.printer,
-                    RepetierCommandFeature.list)
-                    .ConfigureAwait(false);
-                */
-                RepetierPrinterListRespone respone = GetObjectFromJson<RepetierPrinterListRespone>(result.Result);
-                if (respone != null)
-                {                   
-                    repetierPrinterList = new ObservableCollection<IPrinter3d>(respone.Printers);
-                    foreach (RepetierPrinter printer in repetierPrinterList.Cast<RepetierPrinter>())
-                    {
-                        if (printer?.JobId > 0)
-                        {
-                            IPrinter3d prevPrinter = Printers?.FirstOrDefault(p => p.Slug == printer.Slug);
-                            if (prevPrinter is null) continue;
-                            // Avoid unnecessary calls if the image or the job hasn't changed
-                            if (prevPrinter?.ActiveJobId != printer?.ActiveJobId || prevPrinter?.CurrentPrintImage?.Length <= 0)
-                            {
-                                printer.CurrentPrintImage = await GetDynamicRenderImageByJobIdAsync(printer.JobId, false).ConfigureAwait(false);
-                            }
-                            else
-                            {
-                                printer.CurrentPrintImage = prevPrinter.CurrentPrintImage;
-                            }
-                        }
-                        else printer.CurrentPrintImage = Array.Empty<byte>();
-                    }
-                    Printers = repetierPrinterList;
-                }
-
-                return repetierPrinterList;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(RepetierPrinter),
-                    Message = jecx.Message,
-                });
-                return new ObservableCollection<IPrinter3d>();
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return new ObservableCollection<IPrinter3d>();
-            }
-        }
-        public async Task RefreshPrinterListAsync()
-        {
-            try
-            {
-                ObservableCollection<IPrinter3d> printers = new();
-                if (!IsReady)
-                {
-                    Printers = printers;
-                    return;
-                }
-
-                ObservableCollection<IPrinter3d> result = await GetPrintersAsync().ConfigureAwait(false);
-                if (result != null)
-                {
-                    Printers = result;
-                }
-                else
-                {
-                    Printers = printers;
-                }
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                Printers = new ObservableCollection<IPrinter3d>();
-            }
-        }
-        #endregion
-
         #region CurrentPrintInfo
         public async Task<RepetierCurrentPrintInfo> GetCurrentPrintInfoAsync()
         {
@@ -2542,109 +1792,6 @@ namespace AndreasReitberger.API.Repetier
             }
         }
 
-        #endregion
-
-        #region PrinterConfiguration
-        public async Task<RepetierPrinterConfig> GetPrinterConfigAsync(string printerName = "")
-        {
-            RepetierPrinterConfig resultObject = null;
-
-            string currentPrinter = string.IsNullOrEmpty(printerName) ? GetActivePrinterSlug() : printerName;
-            if (string.IsNullOrEmpty(currentPrinter)) return resultObject;
-
-            IRestApiRequestRespone result = null;
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{currentPrinter}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "getPrinterConfig",
-                       jsonObject: new { printer = currentPrinter},
-                       authHeaders: AuthHeaders
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result = await SendRestApiRequestAsync(
-                    RepetierCommandBase.printer, RepetierCommandFeature.api,
-                    command: "getPrinterConfig", jsonData: string.Format("{{\"printer\": \"{0}\"}}", currentPrinter),
-                    printerName: currentPrinter)
-                    .ConfigureAwait(false);
-                */
-                RepetierPrinterConfig config = GetObjectFromJson<RepetierPrinterConfig>(result.Result);
-                if (config != null)
-                {
-                    Config = resultObject = config;
-                }
-                return resultObject;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshPrinterConfigAsync()
-        {
-            try
-            {
-                RepetierPrinterConfig result = await GetPrinterConfigAsync().ConfigureAwait(false);
-                if (result != null)
-                {
-                    Config = result;
-                }
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-            }
-        }
-
-        public async Task<bool> SetPrinterConfigAsync(RepetierPrinterConfig newConfig, string printerName = "")
-        {
-            string currentPrinter = string.IsNullOrEmpty(printerName) ? GetActivePrinterSlug() : printerName;
-            if (string.IsNullOrEmpty(currentPrinter))
-            {
-                return false;
-            }
-
-            IRestApiRequestRespone result = null;
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{currentPrinter}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "setPrinterConfig",
-                       jsonObject: newConfig,
-                       authHeaders: AuthHeaders
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result = await SendRestApiRequestAsync(
-                    RepetierCommandBase.printer, RepetierCommandFeature.api,
-                    command: "setPrinterConfig", jsonData: newConfig,
-                    printerName: currentPrinter
-                    ).ConfigureAwait(false);
-                */
-                return GetQueryResult(result?.Result, true);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
         #endregion
 
         #region Control Commands
